api/requirements.txt
fastapi
uvicorn
requests
pyjwt
httpx
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
